friends

+++ lib/features/profile/data/repository/logout_repo.dart
@@ -1,23 +1,31 @@
-import 'package:cadeau/core/data/apis/api_consumer.dart';
-import 'package:cadeau/core/data/apis/app_endpoint.dart';
-import 'package:cadeau/core/data/error/errorModel.dart';
-import 'package:cadeau/core/data/error/serverException.dart';
+import 'package:cadeau/core/constant/app_endpoint.dart';
+import 'package:cadeau/core/errors/exceptions.dart';
+import 'package:cadeau/core/errors/failures.dart';
+
+import 'package:cadeau/core/network/apis/api_consumer.dart';
+import 'package:cadeau/core/network/network_info.dart';
 import 'package:cadeau/features/profile/data/model/logout_model.dart';
 import 'package:dartz/dartz.dart';
 
 class LogoutRepo {
   final ApiConsumer api;
-  LogoutRepo({required this.api});
+  final NetworkInfo networkInfo;
 
-  Future<Either<ErrorModel, LogoutModel>> logoutRequest() async {
-    try {
-      final response = await api.delete(AppEndpoint.logout, withToken: true);
-      final logoutUser = LogoutModel.fromJson(response);
-      return Right(logoutUser);
-    } on ServerException catch (e) {
-      return Left(e.errModel);
-    } catch (e) {
-      return Left(ErrorModel(status: false, message: e.toString()));
+  LogoutRepo({required this.api, required this.networkInfo});
+
+  Future<Either<Failure, LogoutModel>> logoutRequest() async {
+    if (await networkInfo.isConnected) {
+      try {
+        final response = await api.delete(AppEndpoint.logout, withToken: true);
+        final logoutUser = LogoutModel.fromJson(response);
+        return Right(logoutUser);
+      } on ServerException catch (e) {
+        return Left(ServerFailure(e.message, statusCode: e.statusCode));
+      } catch (e) {
+        return Left(ServerFailure('😭😭Error in: ${e.toString()}'));
+      }
+    } else {
+      return Left(NetworkFailure());
     }
   }
 }
